refactor: 관련있는 dto끼리 내부 클래스로 묶어서 관리
