--- src/main/java/service/dto/AttendanceStatusCount.java
@@ -1,9 +1,0 @@
-package service.dto;
-
-public record AttendanceStatusCount(
-        int attendanceCount,
-        int lateCount,
-        int absentCount
-) {
-    
-}
