--- src/main/java/service/dto/RiskCrew.java
@@ -1,12 +1,0 @@
-package service.dto;
-
-import domain.RiskRank;
-
-public record RiskCrew(
-        String nickname,
-        int lateCount,
-        int absentCount,
-        RiskRank riskRank
-) {
-    
-}
